+++ Aelia's Apprentice/Assets/Scripts/HitResponder.cs
@@ -7,8 +7,12 @@ public class HitResponder : MonoBehaviour
     private Hitbox Hitbox { get; set; }
 
     // Determines if a hitresponder damages
-    [SerializeField] private bool doesDamage = false;
+    [SerializeField] private bool damages = false;
+    [SerializeField] private bool freezes = false;
+    [SerializeField] private bool moves = false;
+    [SerializeField] private bool resizes = false;
 
+    // Only call and override if it is a hitbox responder
     protected virtual void Start()
     {
         Hitbox = GetComponentInChildren<Hitbox>();
@@ -22,15 +26,43 @@ public class HitResponder : MonoBehaviour
 
     private void OnHit(Collider2D hitboxCollider)
     {
-        Debug.Log(gameObject + ": My hitbox has hit something");
+        HandleHit(hitboxCollider);
+    }
 
-        // If the hitbox collider object is damageable, have it take damage
-        IDamageable damageable = hitboxCollider.GetComponentInParent<IDamageable>();
-        if (damageable != null && doesDamage)
+    protected void HandleHit(Collider2D hitboxCollider)
+    {
+        Debug.Log(gameObject + ": I have hit something");
+
+        // If this can damage, attempt to damage the target
+        if (damages)
         {
-            damageable.TakeDamage();
+            IDamageable damageable = hitboxCollider.GetComponentInParent<IDamageable>();
+            if (damageable != null) damageable.TakeDamage();
+            else Debug.Log(hitboxCollider.gameObject + " was not damaged");
         }
-        else
-            Debug.Log(hitboxCollider.gameObject + " was not damaged");
+
+        // If this can freeze, attempt to freeze the target
+        if (freezes)
+        {
+            IFreezeable freezeable = hitboxCollider.GetComponentInParent<IFreezeable>();
+            if (freezeable != null) freezeable.Freeze();
+            else Debug.Log(hitboxCollider.gameObject + " was not frozen");
+        }
+
+        // If this can move, attempt to move the target
+        if (moves)
+        {
+            IMoveable moveable = hitboxCollider.GetComponentInParent<IMoveable>();
+            if (moveable != null) moveable.ForcedMove();
+            else Debug.Log(hitboxCollider.gameObject + " was not moved");
+        }
+
+        // If this can resize, attempt to resize the target
+        if (resizes)
+        {
+            ISizeable sizeable = hitboxCollider.GetComponentInParent<ISizeable>();
+            if (sizeable != null) sizeable.ChangeSize();
+            else Debug.Log(hitboxCollider.gameObject + " was not resized");
+        }
     }
 }