+++ Aelia's Apprentice/Assets/Scripts/IFreezeable.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public interface IFreezeable
+{
+    void Freeze();
+}